+++ app/src/main/java/com/task/noteapp/feature_folder/domain/repository/FolderRepository.kt
@@ -6,7 +6,7 @@ import io.reactivex.rxjava3.core.Maybe
 
 interface FolderRepository {
     fun getAllFolders(): Maybe<Folder>
-    fun getFolderById(id: Long): Maybe<Folder>
+    fun getFolderById(id: Int): Maybe<Folder>
     fun createFolder(folder: Folder): Completable
     fun deleteFolder(folder: Folder): Completable
     fun deleteAllFolders(): Completable
